requirements.txt
celery==5.4.0
channels==4.0.0
daphne==4.1.2
gevent==24.10.3
Django==5.1.2
django-celery-beat==2.7.0
django-cors-headers==4.5.0
django_parler==2.3
django_parler_rest==2.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
python-decouple==3.8
psycopg==3.2
redis==5.1.1
pillow==11.0.0
kavenegar==1.1.2